Changed some placeholder trips and added reviews to cover multiple use cases
Trip 1 is now an example of completed trip which our logged in user (id = 1) has taken part and already leaved reviews.
Trip 2 is now an example of completed trip where the logged in user has taken part but not leaved reviews yet.
Trip 3 is now an example of a trip that has yet to start and the logged in user is already a participant.
Trip 6 is now an example where the logged in use has alredy sent a request to join (for 1 spot).